test(zeroshot-rust): split Codex correction coverage
